Encapsulate UI logic in new module, ModuleUI
