+++ Engine/Application.cpp
@@ -9,9 +9,9 @@
 #include "Module/ModuleScene.h"
 #include "Module/ModuleTexture.h"
 #include "Module/ModuleTime.h"
+#include "Module/ModuleUI.h"
 #include "Module/ModuleFileSystem.h"
 #include "Module/ModuleWindow.h"
-#include "UI/EngineUI.h"
 #include "UI/EngineLog.h"
 #include "TimerUs.h"
 #include "Brofiler/Brofiler.h"
@@ -25,6 +25,7 @@ Application::Application()
 	modules.emplace_back(window = new ModuleWindow());
 	modules.emplace_back(filesystem = new ModuleFileSystem());
 	modules.emplace_back(editor = new ModuleEditor());
+	modules.emplace_back(ui = new ModuleUI());
 	modules.emplace_back(time = new ModuleTime());
 	modules.emplace_back(texture = new ModuleTexture());
 	modules.emplace_back(renderer = new ModuleRender());
@@ -36,7 +37,6 @@ Application::Application()
 	modules.emplace_back(debug = new ModuleDebug());
 		
 	engine_log = new EngineLog();
-	ui = new EngineUI();
 }
 
 Application::~Application()
@@ -46,7 +46,6 @@ Application::~Application()
         delete *it;
     }
 
-	delete ui;
 	delete engine_log;
 }
