+++ Engine/Module/ModuleEditor.h
@@ -4,8 +4,6 @@
 #include "Module.h"
 #include "Globals.h"
 
-struct ImFont;
-
 class ModuleEditor : public Module
 {
 public:
@@ -16,13 +14,8 @@ public:
 	update_status PreUpdate();
 	update_status Update();
 	update_status PostUpdate();
-	void Render();
 	bool CleanUp();
 
-	ImFont* GetFont(const Fonts font) const;
-
-private:
-	void LoadFonts();
 };
 
 #endif //_MODULEEDITOR_H_
